requirements.txt
streamlit
opencv-python
face_recognition
yagmail
geocoder
streamlit-webrtc
pillow
